requirements.txt
betamax==0.8.1
dj_database_url==1.2.0
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
django-on-heroku==1.1.2
djoser==2.2.0
environ==1.0
gunicorn==21.2.0
itemadapter==0.7.0
psycopg2==2.9.7
psycopg2-binary==2.9.5
python-dotenv==0.21.1
python-scrapyd-api==2.1.2
Scrapy==2.8.0
scrapyd==1.4.1
scrapyd-heroku==0.2.2
scrapyd_api==0.0.1
django-prometheus==2.3.1
django-health-check==3.17.0
msgpack==1.0.7
scrapinghub==2.4.0